requirements.txt
fastapi
uvicorn
kubernetes
pydantic
PyYAML
websockets
livekit
